+++ .formatter.exs
@@ -1,5 +1,4 @@
 # Used by "mix format"
 [
-  import_deps: [:mneme],
   inputs: ["{mix,.formatter,.credo}.exs", "{config,lib,test}/**/*.{ex,exs}"]
 ]